Decide to use the ready-to-use ros_gazebo plugins instead of hand written old ones, add slam config files
