+++ dustbuster_ws/launch/gazebo(legacy)/models/room_walls/model.config
@@ -1,0 +1,11 @@
+<?xml version="1.0" ?>
+<model>
+    <name>room_walls</name>
+    <version>1.0</version>
+    <sdf version="1.7">model.sdf</sdf>
+    <author>
+        <name></name>
+        <email></email>
+    </author>
+    <description></description>
+</model>
