--- dustbuster_ws/launch/gazebo/models/lidar_2d/model.config
@@ -1,11 +1,0 @@
-<?xml version="1.0" ?>
-<model>
-    <name>lidar_2d</name>
-    <version>1.0</version>
-    <sdf version="1.7">model.sdf</sdf>
-    <author>
-        <name></name>
-        <email></email>
-    </author>
-    <description></description>
-</model>
