--- dustbuster_ws/launch/gazebo/models/robot_draft/model.config
@@ -1,11 +1,0 @@
-<?xml version="1.0" ?>
-<model>
-    <name>robot_draft</name>
-    <version>1.0</version>
-    <sdf version="1.7">model.sdf</sdf>
-    <author>
-        <name></name>
-        <email></email>
-    </author>
-    <description></description>
-</model>
