--- dustbuster_ws/launch/gazebo/models/room_walls/model.config
@@ -1,11 +1,0 @@
-<?xml version="1.0" ?>
-<model>
-    <name>room_walls</name>
-    <version>1.0</version>
-    <sdf version="1.7">model.sdf</sdf>
-    <author>
-        <name></name>
-        <email></email>
-    </author>
-    <description></description>
-</model>
