--- dustbuster_ws/launch/gazebo/models/room_walls/model.sdf
@@ -1,210 +1,0 @@
-<?xml version='1.0'?>
-<sdf version='1.7'>
-  <model name='room_walls'>
-    <pose>-1.084 -0.115 0 0 -0 0</pose>
-    <link name='Wall_1'>
-      <collision name='Wall_1_Collision'>
-        <geometry>
-          <box>
-            <size>4.5 0.15 1</size>
-          </box>
-        </geometry>
-        <pose>0 0 0.5 0 -0 0</pose>
-      </collision>
-      <visual name='Wall_1_Visual'>
-        <pose>0 0 0.5 0 -0 0</pose>
-        <geometry>
-          <box>
-            <size>4.5 0.15 1</size>
-          </box>
-        </geometry>
-        <material>
-          <script>
-            <uri>file://media/materials/scripts/gazebo.material</uri>
-            <name>Gazebo/Grey</name>
-          </script>
-          <ambient>1 1 1 1</ambient>
-        </material>
-        <meta>
-          <layer>0</layer>
-        </meta>
-      </visual>
-      <pose>-4.425 -4.925 0 0 -0 3.14159</pose>
-    </link>
-    <link name='Wall_2'>
-      <collision name='Wall_2_Collision'>
-        <geometry>
-          <box>
-            <size>10 0.15 1</size>
-          </box>
-        </geometry>
-        <pose>0 0 0.5 0 -0 0</pose>
-      </collision>
-      <visual name='Wall_2_Visual'>
-        <pose>0 0 0.5 0 -0 0</pose>
-        <geometry>
-          <box>
-            <size>10 0.15 1</size>
-          </box>
-        </geometry>
-        <material>
-          <script>
-            <uri>file://media/materials/scripts/gazebo.material</uri>
-            <name>Gazebo/Grey</name>
-          </script>
-          <ambient>1 1 1 1</ambient>
-        </material>
-        <meta>
-          <layer>0</layer>
-        </meta>
-      </visual>
-      <pose>-6.6 -0 0 0 -0 1.5708</pose>
-    </link>
-    <link name='Wall_3'>
-      <collision name='Wall_3_Collision'>
-        <geometry>
-          <box>
-            <size>11.5 0.15 1</size>
-          </box>
-        </geometry>
-        <pose>0 0 0.5 0 -0 0</pose>
-      </collision>
-      <visual name='Wall_3_Visual'>
-        <pose>0 0 0.5 0 -0 0</pose>
-        <geometry>
-          <box>
-            <size>11.5 0.15 1</size>
-          </box>
-        </geometry>
-        <material>
-          <script>
-            <uri>file://media/materials/scripts/gazebo.material</uri>
-            <name>Gazebo/Grey</name>
-          </script>
-          <ambient>1 1 1 1</ambient>
-        </material>
-        <meta>
-          <layer>0</layer>
-        </meta>
-      </visual>
-      <pose>-0.925 4.925 0 0 -0 0</pose>
-    </link>
-    <link name='Wall_4'>
-      <collision name='Wall_4_Collision'>
-        <geometry>
-          <box>
-            <size>1.75 0.15 1</size>
-          </box>
-        </geometry>
-        <pose>0 0 0.5 0 -0 0</pose>
-      </collision>
-      <visual name='Wall_4_Visual'>
-        <pose>0 0 0.5 0 -0 0</pose>
-        <geometry>
-          <box>
-            <size>1.75 0.15 1</size>
-          </box>
-        </geometry>
-        <material>
-          <script>
-            <uri>file://media/materials/scripts/gazebo.material</uri>
-            <name>Gazebo/Grey</name>
-          </script>
-          <ambient>1 1 1 1</ambient>
-        </material>
-        <meta>
-          <layer>0</layer>
-        </meta>
-      </visual>
-      <pose>4.75 4.125 0 0 -0 -1.5708</pose>
-    </link>
-    <link name='Wall_5'>
-      <collision name='Wall_5_Collision'>
-        <geometry>
-          <box>
-            <size>2 0.15 1</size>
-          </box>
-        </geometry>
-        <pose>0 0 0.5 0 -0 0</pose>
-      </collision>
-      <visual name='Wall_5_Visual'>
-        <pose>0 0 0.5 0 -0 0</pose>
-        <geometry>
-          <box>
-            <size>2 0.15 1</size>
-          </box>
-        </geometry>
-        <material>
-          <script>
-            <uri>file://media/materials/scripts/gazebo.material</uri>
-            <name>Gazebo/Grey</name>
-          </script>
-          <ambient>1 1 1 1</ambient>
-        </material>
-        <meta>
-          <layer>0</layer>
-        </meta>
-      </visual>
-      <pose>5.675 3.325 0 0 -0 0</pose>
-    </link>
-    <link name='Wall_6'>
-      <collision name='Wall_6_Collision'>
-        <geometry>
-          <box>
-            <size>8.4 0.15 1</size>
-          </box>
-        </geometry>
-        <pose>0 0 0.5 0 -0 0</pose>
-      </collision>
-      <visual name='Wall_6_Visual'>
-        <pose>0 0 0.5 0 -0 0</pose>
-        <geometry>
-          <box>
-            <size>8.4 0.15 1</size>
-          </box>
-        </geometry>
-        <material>
-          <script>
-            <uri>file://media/materials/scripts/gazebo.material</uri>
-            <name>Gazebo/Grey</name>
-          </script>
-          <ambient>1 1 1 1</ambient>
-        </material>
-        <meta>
-          <layer>0</layer>
-        </meta>
-      </visual>
-      <pose>6.6 -0.8 0 0 -0 -1.5708</pose>
-    </link>
-    <link name='Wall_9'>
-      <collision name='Wall_9_Collision'>
-        <geometry>
-          <box>
-            <size>6.5 0.15 1</size>
-          </box>
-        </geometry>
-        <pose>0 0 0.5 0 -0 0</pose>
-      </collision>
-      <visual name='Wall_9_Visual'>
-        <pose>0 0 0.5 0 -0 0</pose>
-        <geometry>
-          <box>
-            <size>6.5 0.15 1</size>
-          </box>
-        </geometry>
-        <material>
-          <script>
-            <uri>file://media/materials/scripts/gazebo.material</uri>
-            <name>Gazebo/Grey</name>
-          </script>
-          <ambient>1 1 1 1</ambient>
-        </material>
-        <meta>
-          <layer>0</layer>
-        </meta>
-      </visual>
-      <pose>3.425 -4.925 0 0 -0 3.14159</pose>
-    </link>
-    <static>1</static>
-  </model>
-</sdf>
