--- dustbuster_ws/launch/gazebo/plugins/package.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>db_gazebo_plugins</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="onurulusoys4@gmail.com">onur</maintainer>
-  <license>TODO: License declaration</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>gazebo_ros_pkgs</depend>
-  <depend>rclcpp</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
