+++ dustbuster_ws/launch/slam_toolbox_config.yaml
@@ -1,0 +1,14 @@
+slam_toolbox:
+  ros__parameters:
+    odom_topic: /odom
+    imu_topic: /imu
+    use_sim_time: False
+    mode: "mapping"
+    map_update_interval: 3.0
+    max_laser_range: 30.0
+    min_laser_range: 0.1
+    resolution: 0.05
+    scan_topic: /scan
+    use_imu: true
+    use_odom: true
+
